+++ project/home/templates/about.html
@@ -1,6 +1,6 @@
 {% extends "layout.html" %}
 {% block content %}
 
-home
+about
 
 {% endblock %}
