+++ project/home/templates/classify.html
@@ -1,0 +1,23 @@
+{% extends "layout.html" %}
+{% block content %}
+
+<div class="photo-container">
+  <img src='/static/images/placeholder.jpg'>
+</div>
+
+<div class="menu-container">
+  <h2>Classify</h2>
+
+  <h3>Features</h3>
+
+  <ul>
+    <li><a href="#">layered rocks</a></li>
+    <li><a href="#">gypsum vein</a></li>
+    <li><a href="#">bright rocks</a></li>
+    <li><a href="#">meteorites</a></li>
+    <li><a href="#">blueberries</a></li>
+    <li><a href="#">wheel wear</a></li>
+  </ul>
+</div>
+
+{% endblock %}
