+++ project/home/templates/stats.html
@@ -1,6 +1,6 @@
 {% extends "layout.html" %}
 {% block content %}
 
-home
+stat
 
 {% endblock %}
